pyproject.toml
[build-system]
requires = ["setuptools>=42"]
build-backend = "setuptools.build_meta"

[project]
name = "pdf_bookmark_automator"
version = "1.0.0"
description = "Automate PDF bookmark creation using OCR and AI."
authors = [
    { name="Your Name", email="your.email@example.com" }
]
license = { file="LICENSE" }
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "customtkinter>=5.2.0",
    "pymupdf>=1.23.5",
    "pytesseract>=0.3.10",
    "pillow>=10.0.0",
    "openai>=0.28.0",
    "pyperclip>=1.8.2"
]